feat(plugin-authentication): administration redesign
CB-259
